@@ -1,4 +1,4 @@
-// swift-tools-version: 5.8
+// swift-tools-version: 5.10
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -15,7 +15,7 @@ let package = Package(
     ],
     dependencies: [
         .package(url: "https://github.com/apple/swift-protobuf.git", from: "1.25.2"),
-        .package(url: "https://github.com/grpc/grpc-swift.git", branch: "main"),
+        .package(url: "https://github.com/grpc/grpc-swift.git", from: "1.24.1"),
         .package(url: "https://github.com/apple/swift-log.git", branch: "main")
     ],
     targets: [
