@@ -1,5 +1,6 @@
 // DO NOT EDIT.
 // swift-format-ignore-file
+// swiftlint:disable all
 //
 // Generated by the Swift generator plugin for the protocol buffer compiler.
 // Source: google/api/resource.proto
@@ -21,7 +22,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import Foundation
 import SwiftProtobuf
 
 // If the compiler emits an error on this type, it is because this file
@@ -81,7 +81,7 @@ fileprivate struct _GeneratedWithProtocGenSwiftVersion: SwiftProtobuf.ProtobufAP
 ///       pattern: "folders/{folder}/logs/{log}"
 ///       pattern: "organizations/{organization}/logs/{log}"
 ///       pattern: "billingAccounts/{billing_account}/logs/{log}"
-public struct Google_Api_ResourceDescriptor {
+public struct Google_Api_ResourceDescriptor: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -169,7 +169,7 @@ public struct Google_Api_ResourceDescriptor {
 
   /// A description of the historical or future-looking state of the
   /// resource pattern.
-  public enum History: SwiftProtobuf.Enum {
+  public enum History: SwiftProtobuf.Enum, Swift.CaseIterable {
     public typealias RawValue = Int
 
     /// The "unset" value.
@@ -207,10 +207,17 @@ public struct Google_Api_ResourceDescriptor {
       }
     }
 
+    // The compiler won't synthesize support with the UNRECOGNIZED case.
+    public static let allCases: [Google_Api_ResourceDescriptor.History] = [
+      .unspecified,
+      .originallySinglePattern,
+      .futureMultiPattern,
+    ]
+
   }
 
   /// A flag representing a specific style that a resource claims to conform to.
-  public enum Style: SwiftProtobuf.Enum {
+  public enum Style: SwiftProtobuf.Enum, Swift.CaseIterable {
     public typealias RawValue = Int
 
     /// The unspecified value. Do not use.
@@ -247,35 +254,20 @@ public struct Google_Api_ResourceDescriptor {
       }
     }
 
+    // The compiler won't synthesize support with the UNRECOGNIZED case.
+    public static let allCases: [Google_Api_ResourceDescriptor.Style] = [
+      .unspecified,
+      .declarativeFriendly,
+    ]
+
   }
 
   public init() {}
 }
 
-#if swift(>=4.2)
-
-extension Google_Api_ResourceDescriptor.History: CaseIterable {
-  // The compiler won't synthesize support with the UNRECOGNIZED case.
-  public static let allCases: [Google_Api_ResourceDescriptor.History] = [
-    .unspecified,
-    .originallySinglePattern,
-    .futureMultiPattern,
-  ]
-}
-
-extension Google_Api_ResourceDescriptor.Style: CaseIterable {
-  // The compiler won't synthesize support with the UNRECOGNIZED case.
-  public static let allCases: [Google_Api_ResourceDescriptor.Style] = [
-    .unspecified,
-    .declarativeFriendly,
-  ]
-}
-
-#endif  // swift(>=4.2)
-
 /// Defines a proto annotation that describes a string field that refers to
 /// an API resource.
-public struct Google_Api_ResourceReference {
+public struct Google_Api_ResourceReference: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -320,18 +312,11 @@ public struct Google_Api_ResourceReference {
   public init() {}
 }
 
-#if swift(>=5.5) && canImport(_Concurrency)
-extension Google_Api_ResourceDescriptor: @unchecked Sendable {}
-extension Google_Api_ResourceDescriptor.History: @unchecked Sendable {}
-extension Google_Api_ResourceDescriptor.Style: @unchecked Sendable {}
-extension Google_Api_ResourceReference: @unchecked Sendable {}
-#endif  // swift(>=5.5) && canImport(_Concurrency)
-
 // MARK: - Extension support defined in resource.proto.
 
 // MARK: - Extension Properties
 
-// Swift Extensions on the exteneded Messages to add easy access to the declared
+// Swift Extensions on the extended Messages to add easy access to the declared
 // extension fields. The names are based on the extension field name from the proto
 // declaration. To avoid naming collisions, the names are prefixed with the name of
 // the scope where the extend directive occurs.
@@ -363,16 +348,6 @@ extension SwiftProtobuf.Google_Protobuf_FileOptions {
   public var Google_Api_resourceDefinition: [Google_Api_ResourceDescriptor] {
     get {return getExtensionValue(ext: Google_Api_Extensions_resource_definition) ?? []}
     set {setExtensionValue(ext: Google_Api_Extensions_resource_definition, value: newValue)}
-  }
-  /// Returns true if extension `Google_Api_Extensions_resource_definition`
-  /// has been explicitly set.
-  public var hasGoogle_Api_resourceDefinition: Bool {
-    return hasExtensionValue(ext: Google_Api_Extensions_resource_definition)
-  }
-  /// Clears the value of extension `Google_Api_Extensions_resource_definition`.
-  /// Subsequent reads from it will return its default value.
-  public mutating func clearGoogle_Api_resourceDefinition() {
-    clearExtensionValue(ext: Google_Api_Extensions_resource_definition)
   }
 }
 
