@@ -1,5 +1,6 @@
 // DO NOT EDIT.
 // swift-format-ignore-file
+// swiftlint:disable all
 //
 // Generated by the Swift generator plugin for the protocol buffer compiler.
 // Source: google/firestore/v1/bloom_filter.proto
@@ -49,7 +50,7 @@ fileprivate struct _GeneratedWithProtocGenSwiftVersion: SwiftProtobuf.ProtobufAP
 ///
 /// The "size" of a `BitSequence` (the number of bits it contains) is calculated
 /// by this formula: `(bitmap.length * 8) - padding`.
-public struct Google_Firestore_V1_BitSequence {
+public struct Google_Firestore_V1_BitSequence: @unchecked Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -81,7 +82,7 @@ public struct Google_Firestore_V1_BitSequence {
 ///
 /// These resulting values are then taken modulo the number of bits in the bloom
 /// filter to get the bits of the bloom filter to test for the given entry.
-public struct Google_Firestore_V1_BloomFilter {
+public struct Google_Firestore_V1_BloomFilter: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -105,11 +106,6 @@ public struct Google_Firestore_V1_BloomFilter {
 
   fileprivate var _bits: Google_Firestore_V1_BitSequence? = nil
 }
-
-#if swift(>=5.5) && canImport(_Concurrency)
-extension Google_Firestore_V1_BitSequence: @unchecked Sendable {}
-extension Google_Firestore_V1_BloomFilter: @unchecked Sendable {}
-#endif  // swift(>=5.5) && canImport(_Concurrency)
 
 // MARK: - Code below here is support for the SwiftProtobuf runtime.
 
