@@ -1,5 +1,6 @@
 // DO NOT EDIT.
 // swift-format-ignore-file
+// swiftlint:disable all
 //
 // Generated by the Swift generator plugin for the protocol buffer compiler.
 // Source: google/firestore/v1/common.proto
@@ -40,7 +41,7 @@ fileprivate struct _GeneratedWithProtocGenSwiftVersion: SwiftProtobuf.ProtobufAP
 /// This is different from standard field masks, as this is always scoped to a
 /// [Document][google.firestore.v1.Document], and takes in account the dynamic
 /// nature of [Value][google.firestore.v1.Value].
-public struct Google_Firestore_V1_DocumentMask {
+public struct Google_Firestore_V1_DocumentMask: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -56,7 +57,7 @@ public struct Google_Firestore_V1_DocumentMask {
 }
 
 /// A precondition on a document, used for conditional operations.
-public struct Google_Firestore_V1_Precondition {
+public struct Google_Firestore_V1_Precondition: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -87,7 +88,7 @@ public struct Google_Firestore_V1_Precondition {
   public var unknownFields = SwiftProtobuf.UnknownStorage()
 
   /// The type of precondition.
-  public enum OneOf_ConditionType: Equatable {
+  public enum OneOf_ConditionType: Equatable, Sendable {
     /// When set to `true`, the target document must exist.
     /// When set to `false`, the target document must not exist.
     case exists(Bool)
@@ -95,31 +96,13 @@ public struct Google_Firestore_V1_Precondition {
     /// that time. Timestamp must be microsecond aligned.
     case updateTime(SwiftProtobuf.Google_Protobuf_Timestamp)
 
-  #if !swift(>=4.1)
-    public static func ==(lhs: Google_Firestore_V1_Precondition.OneOf_ConditionType, rhs: Google_Firestore_V1_Precondition.OneOf_ConditionType) -> Bool {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
-      switch (lhs, rhs) {
-      case (.exists, .exists): return {
-        guard case .exists(let l) = lhs, case .exists(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      case (.updateTime, .updateTime): return {
-        guard case .updateTime(let l) = lhs, case .updateTime(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      default: return false
-      }
-    }
-  #endif
   }
 
   public init() {}
 }
 
 /// Options for creating a new transaction.
-public struct Google_Firestore_V1_TransactionOptions {
+public struct Google_Firestore_V1_TransactionOptions: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -148,37 +131,19 @@ public struct Google_Firestore_V1_TransactionOptions {
   public var unknownFields = SwiftProtobuf.UnknownStorage()
 
   /// The mode of the transaction.
-  public enum OneOf_Mode: Equatable {
+  public enum OneOf_Mode: Equatable, Sendable {
     /// The transaction can only be used for read operations.
     case readOnly(Google_Firestore_V1_TransactionOptions.ReadOnly)
     /// The transaction can be used for both read and write operations.
     case readWrite(Google_Firestore_V1_TransactionOptions.ReadWrite)
 
-  #if !swift(>=4.1)
-    public static func ==(lhs: Google_Firestore_V1_TransactionOptions.OneOf_Mode, rhs: Google_Firestore_V1_TransactionOptions.OneOf_Mode) -> Bool {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
-      switch (lhs, rhs) {
-      case (.readOnly, .readOnly): return {
-        guard case .readOnly(let l) = lhs, case .readOnly(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      case (.readWrite, .readWrite): return {
-        guard case .readWrite(let l) = lhs, case .readWrite(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      default: return false
-      }
-    }
-  #endif
   }
 
   /// Options for a transaction that can be used to read and write documents.
   ///
   /// Firestore does not allow 3rd party auth requests to create read-write.
   /// transactions.
-  public struct ReadWrite {
+  public struct ReadWrite: @unchecked Sendable {
     // SwiftProtobuf.Message conformance is added in an extension below. See the
     // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
     // methods supported on all messages.
@@ -192,7 +157,7 @@ public struct Google_Firestore_V1_TransactionOptions {
   }
 
   /// Options for a transaction that can only be used to read documents.
-  public struct ReadOnly {
+  public struct ReadOnly: Sendable {
     // SwiftProtobuf.Message conformance is added in an extension below. See the
     // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
     // methods supported on all messages.
@@ -218,7 +183,7 @@ public struct Google_Firestore_V1_TransactionOptions {
 
     /// The consistency mode for this transaction. If not set, defaults to strong
     /// consistency.
-    public enum OneOf_ConsistencySelector: Equatable {
+    public enum OneOf_ConsistencySelector: Equatable, Sendable {
       /// Reads documents at the given time.
       ///
       /// This must be a microsecond precision timestamp within the past one
@@ -226,19 +191,6 @@ public struct Google_Firestore_V1_TransactionOptions {
       /// whole minute timestamp within the past 7 days.
       case readTime(SwiftProtobuf.Google_Protobuf_Timestamp)
 
-    #if !swift(>=4.1)
-      public static func ==(lhs: Google_Firestore_V1_TransactionOptions.ReadOnly.OneOf_ConsistencySelector, rhs: Google_Firestore_V1_TransactionOptions.ReadOnly.OneOf_ConsistencySelector) -> Bool {
-        // The use of inline closures is to circumvent an issue where the compiler
-        // allocates stack space for every case branch when no optimizations are
-        // enabled. https://github.com/apple/swift-protobuf/issues/1034
-        switch (lhs, rhs) {
-        case (.readTime, .readTime): return {
-          guard case .readTime(let l) = lhs, case .readTime(let r) = rhs else { preconditionFailure() }
-          return l == r
-        }()
-        }
-      }
-    #endif
     }
 
     public init() {}
@@ -246,17 +198,6 @@ public struct Google_Firestore_V1_TransactionOptions {
 
   public init() {}
 }
-
-#if swift(>=5.5) && canImport(_Concurrency)
-extension Google_Firestore_V1_DocumentMask: @unchecked Sendable {}
-extension Google_Firestore_V1_Precondition: @unchecked Sendable {}
-extension Google_Firestore_V1_Precondition.OneOf_ConditionType: @unchecked Sendable {}
-extension Google_Firestore_V1_TransactionOptions: @unchecked Sendable {}
-extension Google_Firestore_V1_TransactionOptions.OneOf_Mode: @unchecked Sendable {}
-extension Google_Firestore_V1_TransactionOptions.ReadWrite: @unchecked Sendable {}
-extension Google_Firestore_V1_TransactionOptions.ReadOnly: @unchecked Sendable {}
-extension Google_Firestore_V1_TransactionOptions.ReadOnly.OneOf_ConsistencySelector: @unchecked Sendable {}
-#endif  // swift(>=5.5) && canImport(_Concurrency)
 
 // MARK: - Code below here is support for the SwiftProtobuf runtime.
 
