@@ -1,5 +1,6 @@
 // DO NOT EDIT.
 // swift-format-ignore-file
+// swiftlint:disable all
 //
 // Generated by the Swift generator plugin for the protocol buffer compiler.
 // Source: google/firestore/v1/document.proto
@@ -37,7 +38,7 @@ fileprivate struct _GeneratedWithProtocGenSwiftVersion: SwiftProtobuf.ProtobufAP
 /// A Firestore document.
 ///
 /// Must not exceed 1 MiB - 4 bytes.
-public struct Google_Firestore_V1_Document {
+public struct Google_Firestore_V1_Document: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -109,7 +110,7 @@ public struct Google_Firestore_V1_Document {
 }
 
 /// A message that can hold any of the supported value types.
-public struct Google_Firestore_V1_Value {
+public struct Google_Firestore_V1_Value: @unchecked Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -233,7 +234,7 @@ public struct Google_Firestore_V1_Value {
   public var unknownFields = SwiftProtobuf.UnknownStorage()
 
   /// Must have a value set.
-  public enum OneOf_ValueType: Equatable {
+  public enum OneOf_ValueType: Equatable, @unchecked Sendable {
     /// A null value.
     case nullValue(SwiftProtobuf.Google_Protobuf_NullValue)
     /// A boolean value.
@@ -271,67 +272,13 @@ public struct Google_Firestore_V1_Value {
     /// A map value.
     case mapValue(Google_Firestore_V1_MapValue)
 
-  #if !swift(>=4.1)
-    public static func ==(lhs: Google_Firestore_V1_Value.OneOf_ValueType, rhs: Google_Firestore_V1_Value.OneOf_ValueType) -> Bool {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
-      switch (lhs, rhs) {
-      case (.nullValue, .nullValue): return {
-        guard case .nullValue(let l) = lhs, case .nullValue(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      case (.booleanValue, .booleanValue): return {
-        guard case .booleanValue(let l) = lhs, case .booleanValue(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      case (.integerValue, .integerValue): return {
-        guard case .integerValue(let l) = lhs, case .integerValue(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      case (.doubleValue, .doubleValue): return {
-        guard case .doubleValue(let l) = lhs, case .doubleValue(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      case (.timestampValue, .timestampValue): return {
-        guard case .timestampValue(let l) = lhs, case .timestampValue(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      case (.stringValue, .stringValue): return {
-        guard case .stringValue(let l) = lhs, case .stringValue(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      case (.bytesValue, .bytesValue): return {
-        guard case .bytesValue(let l) = lhs, case .bytesValue(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      case (.referenceValue, .referenceValue): return {
-        guard case .referenceValue(let l) = lhs, case .referenceValue(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      case (.geoPointValue, .geoPointValue): return {
-        guard case .geoPointValue(let l) = lhs, case .geoPointValue(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      case (.arrayValue, .arrayValue): return {
-        guard case .arrayValue(let l) = lhs, case .arrayValue(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      case (.mapValue, .mapValue): return {
-        guard case .mapValue(let l) = lhs, case .mapValue(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      default: return false
-      }
-    }
-  #endif
   }
 
   public init() {}
 }
 
 /// An array value.
-public struct Google_Firestore_V1_ArrayValue {
+public struct Google_Firestore_V1_ArrayValue: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -345,7 +292,7 @@ public struct Google_Firestore_V1_ArrayValue {
 }
 
 /// A map value.
-public struct Google_Firestore_V1_MapValue {
+public struct Google_Firestore_V1_MapValue: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -362,14 +309,6 @@ public struct Google_Firestore_V1_MapValue {
 
   public init() {}
 }
-
-#if swift(>=5.5) && canImport(_Concurrency)
-extension Google_Firestore_V1_Document: @unchecked Sendable {}
-extension Google_Firestore_V1_Value: @unchecked Sendable {}
-extension Google_Firestore_V1_Value.OneOf_ValueType: @unchecked Sendable {}
-extension Google_Firestore_V1_ArrayValue: @unchecked Sendable {}
-extension Google_Firestore_V1_MapValue: @unchecked Sendable {}
-#endif  // swift(>=5.5) && canImport(_Concurrency)
 
 // MARK: - Code below here is support for the SwiftProtobuf runtime.
 
