@@ -1,5 +1,6 @@
 // DO NOT EDIT.
 // swift-format-ignore-file
+// swiftlint:disable all
 //
 // Generated by the Swift generator plugin for the protocol buffer compiler.
 // Source: google/firestore/v1/firestore.proto
@@ -36,7 +37,7 @@ fileprivate struct _GeneratedWithProtocGenSwiftVersion: SwiftProtobuf.ProtobufAP
 
 /// The request for
 /// [Firestore.GetDocument][google.firestore.v1.Firestore.GetDocument].
-public struct Google_Firestore_V1_GetDocumentRequest {
+public struct Google_Firestore_V1_GetDocumentRequest: @unchecked Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -88,7 +89,7 @@ public struct Google_Firestore_V1_GetDocumentRequest {
 
   /// The consistency mode for this transaction.
   /// If not set, defaults to strong consistency.
-  public enum OneOf_ConsistencySelector: Equatable {
+  public enum OneOf_ConsistencySelector: Equatable, @unchecked Sendable {
     /// Reads the document in a transaction.
     case transaction(Data)
     /// Reads the version of the document at the given time.
@@ -98,24 +99,6 @@ public struct Google_Firestore_V1_GetDocumentRequest {
     /// minute timestamp within the past 7 days.
     case readTime(SwiftProtobuf.Google_Protobuf_Timestamp)
 
-  #if !swift(>=4.1)
-    public static func ==(lhs: Google_Firestore_V1_GetDocumentRequest.OneOf_ConsistencySelector, rhs: Google_Firestore_V1_GetDocumentRequest.OneOf_ConsistencySelector) -> Bool {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
-      switch (lhs, rhs) {
-      case (.transaction, .transaction): return {
-        guard case .transaction(let l) = lhs, case .transaction(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      case (.readTime, .readTime): return {
-        guard case .readTime(let l) = lhs, case .readTime(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      default: return false
-      }
-    }
-  #endif
   }
 
   public init() {}
@@ -125,7 +108,7 @@ public struct Google_Firestore_V1_GetDocumentRequest {
 
 /// The request for
 /// [Firestore.ListDocuments][google.firestore.v1.Firestore.ListDocuments].
-public struct Google_Firestore_V1_ListDocumentsRequest {
+public struct Google_Firestore_V1_ListDocumentsRequest: @unchecked Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -222,7 +205,7 @@ public struct Google_Firestore_V1_ListDocumentsRequest {
 
   /// The consistency mode for this transaction.
   /// If not set, defaults to strong consistency.
-  public enum OneOf_ConsistencySelector: Equatable {
+  public enum OneOf_ConsistencySelector: Equatable, @unchecked Sendable {
     /// Perform the read as part of an already active transaction.
     case transaction(Data)
     /// Perform the read at the provided time.
@@ -232,24 +215,6 @@ public struct Google_Firestore_V1_ListDocumentsRequest {
     /// minute timestamp within the past 7 days.
     case readTime(SwiftProtobuf.Google_Protobuf_Timestamp)
 
-  #if !swift(>=4.1)
-    public static func ==(lhs: Google_Firestore_V1_ListDocumentsRequest.OneOf_ConsistencySelector, rhs: Google_Firestore_V1_ListDocumentsRequest.OneOf_ConsistencySelector) -> Bool {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
-      switch (lhs, rhs) {
-      case (.transaction, .transaction): return {
-        guard case .transaction(let l) = lhs, case .transaction(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      case (.readTime, .readTime): return {
-        guard case .readTime(let l) = lhs, case .readTime(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      default: return false
-      }
-    }
-  #endif
   }
 
   public init() {}
@@ -259,7 +224,7 @@ public struct Google_Firestore_V1_ListDocumentsRequest {
 
 /// The response for
 /// [Firestore.ListDocuments][google.firestore.v1.Firestore.ListDocuments].
-public struct Google_Firestore_V1_ListDocumentsResponse {
+public struct Google_Firestore_V1_ListDocumentsResponse: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -279,7 +244,7 @@ public struct Google_Firestore_V1_ListDocumentsResponse {
 
 /// The request for
 /// [Firestore.CreateDocument][google.firestore.v1.Firestore.CreateDocument].
-public struct Google_Firestore_V1_CreateDocumentRequest {
+public struct Google_Firestore_V1_CreateDocumentRequest: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -331,7 +296,7 @@ public struct Google_Firestore_V1_CreateDocumentRequest {
 
 /// The request for
 /// [Firestore.UpdateDocument][google.firestore.v1.Firestore.UpdateDocument].
-public struct Google_Firestore_V1_UpdateDocumentRequest {
+public struct Google_Firestore_V1_UpdateDocumentRequest: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -399,7 +364,7 @@ public struct Google_Firestore_V1_UpdateDocumentRequest {
 
 /// The request for
 /// [Firestore.DeleteDocument][google.firestore.v1.Firestore.DeleteDocument].
-public struct Google_Firestore_V1_DeleteDocumentRequest {
+public struct Google_Firestore_V1_DeleteDocumentRequest: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -428,7 +393,7 @@ public struct Google_Firestore_V1_DeleteDocumentRequest {
 
 /// The request for
 /// [Firestore.BatchGetDocuments][google.firestore.v1.Firestore.BatchGetDocuments].
-public struct Google_Firestore_V1_BatchGetDocumentsRequest {
+public struct Google_Firestore_V1_BatchGetDocumentsRequest: @unchecked Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -498,7 +463,7 @@ public struct Google_Firestore_V1_BatchGetDocumentsRequest {
 
   /// The consistency mode for this transaction.
   /// If not set, defaults to strong consistency.
-  public enum OneOf_ConsistencySelector: Equatable {
+  public enum OneOf_ConsistencySelector: Equatable, @unchecked Sendable {
     /// Reads documents in a transaction.
     case transaction(Data)
     /// Starts a new transaction and reads the documents.
@@ -513,28 +478,6 @@ public struct Google_Firestore_V1_BatchGetDocumentsRequest {
     /// minute timestamp within the past 7 days.
     case readTime(SwiftProtobuf.Google_Protobuf_Timestamp)
 
-  #if !swift(>=4.1)
-    public static func ==(lhs: Google_Firestore_V1_BatchGetDocumentsRequest.OneOf_ConsistencySelector, rhs: Google_Firestore_V1_BatchGetDocumentsRequest.OneOf_ConsistencySelector) -> Bool {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
-      switch (lhs, rhs) {
-      case (.transaction, .transaction): return {
-        guard case .transaction(let l) = lhs, case .transaction(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      case (.newTransaction, .newTransaction): return {
-        guard case .newTransaction(let l) = lhs, case .newTransaction(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      case (.readTime, .readTime): return {
-        guard case .readTime(let l) = lhs, case .readTime(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      default: return false
-      }
-    }
-  #endif
   }
 
   public init() {}
@@ -544,7 +487,7 @@ public struct Google_Firestore_V1_BatchGetDocumentsRequest {
 
 /// The streamed response for
 /// [Firestore.BatchGetDocuments][google.firestore.v1.Firestore.BatchGetDocuments].
-public struct Google_Firestore_V1_BatchGetDocumentsResponse {
+public struct Google_Firestore_V1_BatchGetDocumentsResponse: @unchecked Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -595,31 +538,13 @@ public struct Google_Firestore_V1_BatchGetDocumentsResponse {
 
   /// A single result.
   /// This can be empty if the server is just returning a transaction.
-  public enum OneOf_Result: Equatable {
+  public enum OneOf_Result: Equatable, Sendable {
     /// A document that was requested.
     case found(Google_Firestore_V1_Document)
     /// A document name that was requested but does not exist. In the format:
     /// `projects/{project_id}/databases/{database_id}/documents/{document_path}`.
     case missing(String)
 
-  #if !swift(>=4.1)
-    public static func ==(lhs: Google_Firestore_V1_BatchGetDocumentsResponse.OneOf_Result, rhs: Google_Firestore_V1_BatchGetDocumentsResponse.OneOf_Result) -> Bool {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
-      switch (lhs, rhs) {
-      case (.found, .found): return {
-        guard case .found(let l) = lhs, case .found(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      case (.missing, .missing): return {
-        guard case .missing(let l) = lhs, case .missing(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      default: return false
-      }
-    }
-  #endif
   }
 
   public init() {}
@@ -629,7 +554,7 @@ public struct Google_Firestore_V1_BatchGetDocumentsResponse {
 
 /// The request for
 /// [Firestore.BeginTransaction][google.firestore.v1.Firestore.BeginTransaction].
-public struct Google_Firestore_V1_BeginTransactionRequest {
+public struct Google_Firestore_V1_BeginTransactionRequest: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -658,7 +583,7 @@ public struct Google_Firestore_V1_BeginTransactionRequest {
 
 /// The response for
 /// [Firestore.BeginTransaction][google.firestore.v1.Firestore.BeginTransaction].
-public struct Google_Firestore_V1_BeginTransactionResponse {
+public struct Google_Firestore_V1_BeginTransactionResponse: @unchecked Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -672,7 +597,7 @@ public struct Google_Firestore_V1_BeginTransactionResponse {
 }
 
 /// The request for [Firestore.Commit][google.firestore.v1.Firestore.Commit].
-public struct Google_Firestore_V1_CommitRequest {
+public struct Google_Firestore_V1_CommitRequest: @unchecked Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -695,7 +620,7 @@ public struct Google_Firestore_V1_CommitRequest {
 }
 
 /// The response for [Firestore.Commit][google.firestore.v1.Firestore.Commit].
-public struct Google_Firestore_V1_CommitResponse {
+public struct Google_Firestore_V1_CommitResponse: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -725,7 +650,7 @@ public struct Google_Firestore_V1_CommitResponse {
 }
 
 /// The request for [Firestore.Rollback][google.firestore.v1.Firestore.Rollback].
-public struct Google_Firestore_V1_RollbackRequest {
+public struct Google_Firestore_V1_RollbackRequest: @unchecked Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -743,7 +668,7 @@ public struct Google_Firestore_V1_RollbackRequest {
 }
 
 /// The request for [Firestore.RunQuery][google.firestore.v1.Firestore.RunQuery].
-public struct Google_Firestore_V1_RunQueryRequest {
+public struct Google_Firestore_V1_RunQueryRequest: @unchecked Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -822,28 +747,15 @@ public struct Google_Firestore_V1_RunQueryRequest {
   public var unknownFields = SwiftProtobuf.UnknownStorage()
 
   /// The query to run.
-  public enum OneOf_QueryType: Equatable {
+  public enum OneOf_QueryType: Equatable, Sendable {
     /// A structured query.
     case structuredQuery(Google_Firestore_V1_StructuredQuery)
 
-  #if !swift(>=4.1)
-    public static func ==(lhs: Google_Firestore_V1_RunQueryRequest.OneOf_QueryType, rhs: Google_Firestore_V1_RunQueryRequest.OneOf_QueryType) -> Bool {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
-      switch (lhs, rhs) {
-      case (.structuredQuery, .structuredQuery): return {
-        guard case .structuredQuery(let l) = lhs, case .structuredQuery(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      }
-    }
-  #endif
   }
 
   /// The consistency mode for this transaction.
   /// If not set, defaults to strong consistency.
-  public enum OneOf_ConsistencySelector: Equatable {
+  public enum OneOf_ConsistencySelector: Equatable, @unchecked Sendable {
     /// Run the query within an already active transaction.
     ///
     /// The value here is the opaque transaction ID to execute the query in.
@@ -860,28 +772,6 @@ public struct Google_Firestore_V1_RunQueryRequest {
     /// minute timestamp within the past 7 days.
     case readTime(SwiftProtobuf.Google_Protobuf_Timestamp)
 
-  #if !swift(>=4.1)
-    public static func ==(lhs: Google_Firestore_V1_RunQueryRequest.OneOf_ConsistencySelector, rhs: Google_Firestore_V1_RunQueryRequest.OneOf_ConsistencySelector) -> Bool {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
-      switch (lhs, rhs) {
-      case (.transaction, .transaction): return {
-        guard case .transaction(let l) = lhs, case .transaction(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      case (.newTransaction, .newTransaction): return {
-        guard case .newTransaction(let l) = lhs, case .newTransaction(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      case (.readTime, .readTime): return {
-        guard case .readTime(let l) = lhs, case .readTime(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      default: return false
-      }
-    }
-  #endif
   }
 
   public init() {}
@@ -891,7 +781,7 @@ public struct Google_Firestore_V1_RunQueryRequest {
 
 /// The response for
 /// [Firestore.RunQuery][google.firestore.v1.Firestore.RunQuery].
-public struct Google_Firestore_V1_RunQueryResponse {
+public struct Google_Firestore_V1_RunQueryResponse: @unchecked Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -974,24 +864,11 @@ public struct Google_Firestore_V1_RunQueryResponse {
   /// The continuation mode for the query. If present, it indicates the current
   /// query response stream has finished. This can be set with or without a
   /// `document` present, but when set, no more results are returned.
-  public enum OneOf_ContinuationSelector: Equatable {
+  public enum OneOf_ContinuationSelector: Equatable, Sendable {
     /// If present, Firestore has completely finished the request and no more
     /// documents will be returned.
     case done(Bool)
 
-  #if !swift(>=4.1)
-    public static func ==(lhs: Google_Firestore_V1_RunQueryResponse.OneOf_ContinuationSelector, rhs: Google_Firestore_V1_RunQueryResponse.OneOf_ContinuationSelector) -> Bool {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
-      switch (lhs, rhs) {
-      case (.done, .done): return {
-        guard case .done(let l) = lhs, case .done(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      }
-    }
-  #endif
   }
 
   public init() {}
@@ -1001,7 +878,7 @@ public struct Google_Firestore_V1_RunQueryResponse {
 
 /// The request for
 /// [Firestore.RunAggregationQuery][google.firestore.v1.Firestore.RunAggregationQuery].
-public struct Google_Firestore_V1_RunAggregationQueryRequest {
+public struct Google_Firestore_V1_RunAggregationQueryRequest: @unchecked Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -1079,27 +956,14 @@ public struct Google_Firestore_V1_RunAggregationQueryRequest {
   public var unknownFields = SwiftProtobuf.UnknownStorage()
 
   /// The query to run.
-  public enum OneOf_QueryType: Equatable {
+  public enum OneOf_QueryType: Equatable, Sendable {
     /// An aggregation query.
     case structuredAggregationQuery(Google_Firestore_V1_StructuredAggregationQuery)
 
-  #if !swift(>=4.1)
-    public static func ==(lhs: Google_Firestore_V1_RunAggregationQueryRequest.OneOf_QueryType, rhs: Google_Firestore_V1_RunAggregationQueryRequest.OneOf_QueryType) -> Bool {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
-      switch (lhs, rhs) {
-      case (.structuredAggregationQuery, .structuredAggregationQuery): return {
-        guard case .structuredAggregationQuery(let l) = lhs, case .structuredAggregationQuery(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      }
-    }
-  #endif
   }
 
   /// The consistency mode for the query, defaults to strong consistency.
-  public enum OneOf_ConsistencySelector: Equatable {
+  public enum OneOf_ConsistencySelector: Equatable, @unchecked Sendable {
     /// Run the aggregation within an already active transaction.
     ///
     /// The value here is the opaque transaction ID to execute the query in.
@@ -1116,28 +980,6 @@ public struct Google_Firestore_V1_RunAggregationQueryRequest {
     /// minute timestamp within the past 7 days.
     case readTime(SwiftProtobuf.Google_Protobuf_Timestamp)
 
-  #if !swift(>=4.1)
-    public static func ==(lhs: Google_Firestore_V1_RunAggregationQueryRequest.OneOf_ConsistencySelector, rhs: Google_Firestore_V1_RunAggregationQueryRequest.OneOf_ConsistencySelector) -> Bool {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
-      switch (lhs, rhs) {
-      case (.transaction, .transaction): return {
-        guard case .transaction(let l) = lhs, case .transaction(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      case (.newTransaction, .newTransaction): return {
-        guard case .newTransaction(let l) = lhs, case .newTransaction(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      case (.readTime, .readTime): return {
-        guard case .readTime(let l) = lhs, case .readTime(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      default: return false
-      }
-    }
-  #endif
   }
 
   public init() {}
@@ -1147,7 +989,7 @@ public struct Google_Firestore_V1_RunAggregationQueryRequest {
 
 /// The response for
 /// [Firestore.RunAggregationQuery][google.firestore.v1.Firestore.RunAggregationQuery].
-public struct Google_Firestore_V1_RunAggregationQueryResponse {
+public struct Google_Firestore_V1_RunAggregationQueryResponse: @unchecked Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -1210,7 +1052,7 @@ public struct Google_Firestore_V1_RunAggregationQueryResponse {
 
 /// The request for
 /// [Firestore.PartitionQuery][google.firestore.v1.Firestore.PartitionQuery].
-public struct Google_Firestore_V1_PartitionQueryRequest {
+public struct Google_Firestore_V1_PartitionQueryRequest: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -1290,31 +1132,18 @@ public struct Google_Firestore_V1_PartitionQueryRequest {
   public var unknownFields = SwiftProtobuf.UnknownStorage()
 
   /// The query to partition.
-  public enum OneOf_QueryType: Equatable {
+  public enum OneOf_QueryType: Equatable, Sendable {
     /// A structured query.
     /// Query must specify collection with all descendants and be ordered by name
     /// ascending. Other filters, order bys, limits, offsets, and start/end
     /// cursors are not supported.
     case structuredQuery(Google_Firestore_V1_StructuredQuery)
 
-  #if !swift(>=4.1)
-    public static func ==(lhs: Google_Firestore_V1_PartitionQueryRequest.OneOf_QueryType, rhs: Google_Firestore_V1_PartitionQueryRequest.OneOf_QueryType) -> Bool {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
-      switch (lhs, rhs) {
-      case (.structuredQuery, .structuredQuery): return {
-        guard case .structuredQuery(let l) = lhs, case .structuredQuery(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      }
-    }
-  #endif
   }
 
   /// The consistency mode for this request.
   /// If not set, defaults to strong consistency.
-  public enum OneOf_ConsistencySelector: Equatable {
+  public enum OneOf_ConsistencySelector: Equatable, Sendable {
     /// Reads documents as they were at the given time.
     ///
     /// This must be a microsecond precision timestamp within the past one hour,
@@ -1322,19 +1151,6 @@ public struct Google_Firestore_V1_PartitionQueryRequest {
     /// minute timestamp within the past 7 days.
     case readTime(SwiftProtobuf.Google_Protobuf_Timestamp)
 
-  #if !swift(>=4.1)
-    public static func ==(lhs: Google_Firestore_V1_PartitionQueryRequest.OneOf_ConsistencySelector, rhs: Google_Firestore_V1_PartitionQueryRequest.OneOf_ConsistencySelector) -> Bool {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
-      switch (lhs, rhs) {
-      case (.readTime, .readTime): return {
-        guard case .readTime(let l) = lhs, case .readTime(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      }
-    }
-  #endif
   }
 
   public init() {}
@@ -1342,7 +1158,7 @@ public struct Google_Firestore_V1_PartitionQueryRequest {
 
 /// The response for
 /// [Firestore.PartitionQuery][google.firestore.v1.Firestore.PartitionQuery].
-public struct Google_Firestore_V1_PartitionQueryResponse {
+public struct Google_Firestore_V1_PartitionQueryResponse: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -1386,7 +1202,7 @@ public struct Google_Firestore_V1_PartitionQueryResponse {
 /// When resuming a stream, the server first streams any responses later than the
 /// given token, then a response containing only an up-to-date token, to use in
 /// the next request.
-public struct Google_Firestore_V1_WriteRequest {
+public struct Google_Firestore_V1_WriteRequest: @unchecked Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -1434,7 +1250,7 @@ public struct Google_Firestore_V1_WriteRequest {
 }
 
 /// The response for [Firestore.Write][google.firestore.v1.Firestore.Write].
-public struct Google_Firestore_V1_WriteResponse {
+public struct Google_Firestore_V1_WriteResponse: @unchecked Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -1474,7 +1290,7 @@ public struct Google_Firestore_V1_WriteResponse {
 }
 
 /// A request for [Firestore.Listen][google.firestore.v1.Firestore.Listen]
-public struct Google_Firestore_V1_ListenRequest {
+public struct Google_Firestore_V1_ListenRequest: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -1510,37 +1326,19 @@ public struct Google_Firestore_V1_ListenRequest {
   public var unknownFields = SwiftProtobuf.UnknownStorage()
 
   /// The supported target changes.
-  public enum OneOf_TargetChange: Equatable {
+  public enum OneOf_TargetChange: Equatable, Sendable {
     /// A target to add to this stream.
     case addTarget(Google_Firestore_V1_Target)
     /// The ID of a target to remove from this stream.
     case removeTarget(Int32)
 
-  #if !swift(>=4.1)
-    public static func ==(lhs: Google_Firestore_V1_ListenRequest.OneOf_TargetChange, rhs: Google_Firestore_V1_ListenRequest.OneOf_TargetChange) -> Bool {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
-      switch (lhs, rhs) {
-      case (.addTarget, .addTarget): return {
-        guard case .addTarget(let l) = lhs, case .addTarget(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      case (.removeTarget, .removeTarget): return {
-        guard case .removeTarget(let l) = lhs, case .removeTarget(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      default: return false
-      }
-    }
-  #endif
   }
 
   public init() {}
 }
 
 /// The response for [Firestore.Listen][google.firestore.v1.Firestore.Listen].
-public struct Google_Firestore_V1_ListenResponse {
+public struct Google_Firestore_V1_ListenResponse: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -1601,7 +1399,7 @@ public struct Google_Firestore_V1_ListenResponse {
   public var unknownFields = SwiftProtobuf.UnknownStorage()
 
   /// The supported responses.
-  public enum OneOf_ResponseType: Equatable {
+  public enum OneOf_ResponseType: Equatable, Sendable {
     /// Targets have changed.
     case targetChange(Google_Firestore_V1_TargetChange)
     /// A [Document][google.firestore.v1.Document] has changed.
@@ -1618,43 +1416,13 @@ public struct Google_Firestore_V1_ListenResponse {
     /// the exact documents are unknown.
     case filter(Google_Firestore_V1_ExistenceFilter)
 
-  #if !swift(>=4.1)
-    public static func ==(lhs: Google_Firestore_V1_ListenResponse.OneOf_ResponseType, rhs: Google_Firestore_V1_ListenResponse.OneOf_ResponseType) -> Bool {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
-      switch (lhs, rhs) {
-      case (.targetChange, .targetChange): return {
-        guard case .targetChange(let l) = lhs, case .targetChange(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      case (.documentChange, .documentChange): return {
-        guard case .documentChange(let l) = lhs, case .documentChange(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      case (.documentDelete, .documentDelete): return {
-        guard case .documentDelete(let l) = lhs, case .documentDelete(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      case (.documentRemove, .documentRemove): return {
-        guard case .documentRemove(let l) = lhs, case .documentRemove(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      case (.filter, .filter): return {
-        guard case .filter(let l) = lhs, case .filter(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      default: return false
-      }
-    }
-  #endif
   }
 
   public init() {}
 }
 
 /// A specification of a set of documents to listen to.
-public struct Google_Firestore_V1_Target {
+public struct Google_Firestore_V1_Target: @unchecked Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -1750,30 +1518,12 @@ public struct Google_Firestore_V1_Target {
   public var unknownFields = SwiftProtobuf.UnknownStorage()
 
   /// The type of target to listen to.
-  public enum OneOf_TargetType: Equatable {
+  public enum OneOf_TargetType: Equatable, Sendable {
     /// A target specified by a query.
     case query(Google_Firestore_V1_Target.QueryTarget)
     /// A target specified by a set of document names.
     case documents(Google_Firestore_V1_Target.DocumentsTarget)
 
-  #if !swift(>=4.1)
-    public static func ==(lhs: Google_Firestore_V1_Target.OneOf_TargetType, rhs: Google_Firestore_V1_Target.OneOf_TargetType) -> Bool {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
-      switch (lhs, rhs) {
-      case (.query, .query): return {
-        guard case .query(let l) = lhs, case .query(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      case (.documents, .documents): return {
-        guard case .documents(let l) = lhs, case .documents(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      default: return false
-      }
-    }
-  #endif
   }
 
   /// When to start listening.
@@ -1781,7 +1531,7 @@ public struct Google_Firestore_V1_Target {
   /// If specified, only the matching Documents that have been updated AFTER the
   /// `resume_token` or `read_time` will be returned. Otherwise, all matching
   /// Documents are returned before any subsequent changes.
-  public enum OneOf_ResumeType: Equatable {
+  public enum OneOf_ResumeType: Equatable, @unchecked Sendable {
     /// A resume token from a prior
     /// [TargetChange][google.firestore.v1.TargetChange] for an identical target.
     ///
@@ -1792,28 +1542,10 @@ public struct Google_Firestore_V1_Target {
     /// The client must know the state of matching documents at this time.
     case readTime(SwiftProtobuf.Google_Protobuf_Timestamp)
 
-  #if !swift(>=4.1)
-    public static func ==(lhs: Google_Firestore_V1_Target.OneOf_ResumeType, rhs: Google_Firestore_V1_Target.OneOf_ResumeType) -> Bool {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
-      switch (lhs, rhs) {
-      case (.resumeToken, .resumeToken): return {
-        guard case .resumeToken(let l) = lhs, case .resumeToken(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      case (.readTime, .readTime): return {
-        guard case .readTime(let l) = lhs, case .readTime(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      default: return false
-      }
-    }
-  #endif
   }
 
   /// A target specified by a set of documents names.
-  public struct DocumentsTarget {
+  public struct DocumentsTarget: Sendable {
     // SwiftProtobuf.Message conformance is added in an extension below. See the
     // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
     // methods supported on all messages.
@@ -1830,7 +1562,7 @@ public struct Google_Firestore_V1_Target {
   }
 
   /// A target specified by a query.
-  public struct QueryTarget {
+  public struct QueryTarget: Sendable {
     // SwiftProtobuf.Message conformance is added in an extension below. See the
     // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
     // methods supported on all messages.
@@ -1858,23 +1590,10 @@ public struct Google_Firestore_V1_Target {
     public var unknownFields = SwiftProtobuf.UnknownStorage()
 
     /// The query to run.
-    public enum OneOf_QueryType: Equatable {
+    public enum OneOf_QueryType: Equatable, Sendable {
       /// A structured query.
       case structuredQuery(Google_Firestore_V1_StructuredQuery)
 
-    #if !swift(>=4.1)
-      public static func ==(lhs: Google_Firestore_V1_Target.QueryTarget.OneOf_QueryType, rhs: Google_Firestore_V1_Target.QueryTarget.OneOf_QueryType) -> Bool {
-        // The use of inline closures is to circumvent an issue where the compiler
-        // allocates stack space for every case branch when no optimizations are
-        // enabled. https://github.com/apple/swift-protobuf/issues/1034
-        switch (lhs, rhs) {
-        case (.structuredQuery, .structuredQuery): return {
-          guard case .structuredQuery(let l) = lhs, case .structuredQuery(let r) = rhs else { preconditionFailure() }
-          return l == r
-        }()
-        }
-      }
-    #endif
     }
 
     public init() {}
@@ -1886,7 +1605,7 @@ public struct Google_Firestore_V1_Target {
 }
 
 /// Targets being watched have changed.
-public struct Google_Firestore_V1_TargetChange {
+public struct Google_Firestore_V1_TargetChange: @unchecked Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -1939,7 +1658,7 @@ public struct Google_Firestore_V1_TargetChange {
   public var unknownFields = SwiftProtobuf.UnknownStorage()
 
   /// The type of change.
-  public enum TargetChangeType: SwiftProtobuf.Enum {
+  public enum TargetChangeType: SwiftProtobuf.Enum, Swift.CaseIterable {
     public typealias RawValue = Int
 
     /// No change has occurred. Used only to send an updated `resume_token`.
@@ -1995,6 +1714,15 @@ public struct Google_Firestore_V1_TargetChange {
       }
     }
 
+    // The compiler won't synthesize support with the UNRECOGNIZED case.
+    public static let allCases: [Google_Firestore_V1_TargetChange.TargetChangeType] = [
+      .noChange,
+      .add,
+      .remove,
+      .current,
+      .reset,
+    ]
+
   }
 
   public init() {}
@@ -2003,24 +1731,9 @@ public struct Google_Firestore_V1_TargetChange {
   fileprivate var _readTime: SwiftProtobuf.Google_Protobuf_Timestamp? = nil
 }
 
-#if swift(>=4.2)
-
-extension Google_Firestore_V1_TargetChange.TargetChangeType: CaseIterable {
-  // The compiler won't synthesize support with the UNRECOGNIZED case.
-  public static let allCases: [Google_Firestore_V1_TargetChange.TargetChangeType] = [
-    .noChange,
-    .add,
-    .remove,
-    .current,
-    .reset,
-  ]
-}
-
-#endif  // swift(>=4.2)
-
 /// The request for
 /// [Firestore.ListCollectionIds][google.firestore.v1.Firestore.ListCollectionIds].
-public struct Google_Firestore_V1_ListCollectionIdsRequest {
+public struct Google_Firestore_V1_ListCollectionIdsRequest: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -2059,7 +1772,7 @@ public struct Google_Firestore_V1_ListCollectionIdsRequest {
 
   /// The consistency mode for this request.
   /// If not set, defaults to strong consistency.
-  public enum OneOf_ConsistencySelector: Equatable {
+  public enum OneOf_ConsistencySelector: Equatable, Sendable {
     /// Reads documents as they were at the given time.
     ///
     /// This must be a microsecond precision timestamp within the past one hour,
@@ -2067,19 +1780,6 @@ public struct Google_Firestore_V1_ListCollectionIdsRequest {
     /// minute timestamp within the past 7 days.
     case readTime(SwiftProtobuf.Google_Protobuf_Timestamp)
 
-  #if !swift(>=4.1)
-    public static func ==(lhs: Google_Firestore_V1_ListCollectionIdsRequest.OneOf_ConsistencySelector, rhs: Google_Firestore_V1_ListCollectionIdsRequest.OneOf_ConsistencySelector) -> Bool {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
-      switch (lhs, rhs) {
-      case (.readTime, .readTime): return {
-        guard case .readTime(let l) = lhs, case .readTime(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      }
-    }
-  #endif
   }
 
   public init() {}
@@ -2087,7 +1787,7 @@ public struct Google_Firestore_V1_ListCollectionIdsRequest {
 
 /// The response from
 /// [Firestore.ListCollectionIds][google.firestore.v1.Firestore.ListCollectionIds].
-public struct Google_Firestore_V1_ListCollectionIdsResponse {
+public struct Google_Firestore_V1_ListCollectionIdsResponse: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -2105,7 +1805,7 @@ public struct Google_Firestore_V1_ListCollectionIdsResponse {
 
 /// The request for
 /// [Firestore.BatchWrite][google.firestore.v1.Firestore.BatchWrite].
-public struct Google_Firestore_V1_BatchWriteRequest {
+public struct Google_Firestore_V1_BatchWriteRequest: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -2131,7 +1831,7 @@ public struct Google_Firestore_V1_BatchWriteRequest {
 
 /// The response from
 /// [Firestore.BatchWrite][google.firestore.v1.Firestore.BatchWrite].
-public struct Google_Firestore_V1_BatchWriteResponse {
+public struct Google_Firestore_V1_BatchWriteResponse: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -2152,58 +1852,6 @@ public struct Google_Firestore_V1_BatchWriteResponse {
 
   public init() {}
 }
-
-#if swift(>=5.5) && canImport(_Concurrency)
-extension Google_Firestore_V1_GetDocumentRequest: @unchecked Sendable {}
-extension Google_Firestore_V1_GetDocumentRequest.OneOf_ConsistencySelector: @unchecked Sendable {}
-extension Google_Firestore_V1_ListDocumentsRequest: @unchecked Sendable {}
-extension Google_Firestore_V1_ListDocumentsRequest.OneOf_ConsistencySelector: @unchecked Sendable {}
-extension Google_Firestore_V1_ListDocumentsResponse: @unchecked Sendable {}
-extension Google_Firestore_V1_CreateDocumentRequest: @unchecked Sendable {}
-extension Google_Firestore_V1_UpdateDocumentRequest: @unchecked Sendable {}
-extension Google_Firestore_V1_DeleteDocumentRequest: @unchecked Sendable {}
-extension Google_Firestore_V1_BatchGetDocumentsRequest: @unchecked Sendable {}
-extension Google_Firestore_V1_BatchGetDocumentsRequest.OneOf_ConsistencySelector: @unchecked Sendable {}
-extension Google_Firestore_V1_BatchGetDocumentsResponse: @unchecked Sendable {}
-extension Google_Firestore_V1_BatchGetDocumentsResponse.OneOf_Result: @unchecked Sendable {}
-extension Google_Firestore_V1_BeginTransactionRequest: @unchecked Sendable {}
-extension Google_Firestore_V1_BeginTransactionResponse: @unchecked Sendable {}
-extension Google_Firestore_V1_CommitRequest: @unchecked Sendable {}
-extension Google_Firestore_V1_CommitResponse: @unchecked Sendable {}
-extension Google_Firestore_V1_RollbackRequest: @unchecked Sendable {}
-extension Google_Firestore_V1_RunQueryRequest: @unchecked Sendable {}
-extension Google_Firestore_V1_RunQueryRequest.OneOf_QueryType: @unchecked Sendable {}
-extension Google_Firestore_V1_RunQueryRequest.OneOf_ConsistencySelector: @unchecked Sendable {}
-extension Google_Firestore_V1_RunQueryResponse: @unchecked Sendable {}
-extension Google_Firestore_V1_RunQueryResponse.OneOf_ContinuationSelector: @unchecked Sendable {}
-extension Google_Firestore_V1_RunAggregationQueryRequest: @unchecked Sendable {}
-extension Google_Firestore_V1_RunAggregationQueryRequest.OneOf_QueryType: @unchecked Sendable {}
-extension Google_Firestore_V1_RunAggregationQueryRequest.OneOf_ConsistencySelector: @unchecked Sendable {}
-extension Google_Firestore_V1_RunAggregationQueryResponse: @unchecked Sendable {}
-extension Google_Firestore_V1_PartitionQueryRequest: @unchecked Sendable {}
-extension Google_Firestore_V1_PartitionQueryRequest.OneOf_QueryType: @unchecked Sendable {}
-extension Google_Firestore_V1_PartitionQueryRequest.OneOf_ConsistencySelector: @unchecked Sendable {}
-extension Google_Firestore_V1_PartitionQueryResponse: @unchecked Sendable {}
-extension Google_Firestore_V1_WriteRequest: @unchecked Sendable {}
-extension Google_Firestore_V1_WriteResponse: @unchecked Sendable {}
-extension Google_Firestore_V1_ListenRequest: @unchecked Sendable {}
-extension Google_Firestore_V1_ListenRequest.OneOf_TargetChange: @unchecked Sendable {}
-extension Google_Firestore_V1_ListenResponse: @unchecked Sendable {}
-extension Google_Firestore_V1_ListenResponse.OneOf_ResponseType: @unchecked Sendable {}
-extension Google_Firestore_V1_Target: @unchecked Sendable {}
-extension Google_Firestore_V1_Target.OneOf_TargetType: @unchecked Sendable {}
-extension Google_Firestore_V1_Target.OneOf_ResumeType: @unchecked Sendable {}
-extension Google_Firestore_V1_Target.DocumentsTarget: @unchecked Sendable {}
-extension Google_Firestore_V1_Target.QueryTarget: @unchecked Sendable {}
-extension Google_Firestore_V1_Target.QueryTarget.OneOf_QueryType: @unchecked Sendable {}
-extension Google_Firestore_V1_TargetChange: @unchecked Sendable {}
-extension Google_Firestore_V1_TargetChange.TargetChangeType: @unchecked Sendable {}
-extension Google_Firestore_V1_ListCollectionIdsRequest: @unchecked Sendable {}
-extension Google_Firestore_V1_ListCollectionIdsRequest.OneOf_ConsistencySelector: @unchecked Sendable {}
-extension Google_Firestore_V1_ListCollectionIdsResponse: @unchecked Sendable {}
-extension Google_Firestore_V1_BatchWriteRequest: @unchecked Sendable {}
-extension Google_Firestore_V1_BatchWriteResponse: @unchecked Sendable {}
-#endif  // swift(>=5.5) && canImport(_Concurrency)
 
 // MARK: - Code below here is support for the SwiftProtobuf runtime.
 
@@ -3095,7 +2743,15 @@ extension Google_Firestore_V1_RunQueryResponse: SwiftProtobuf.Message, SwiftProt
     var _continuationSelector: Google_Firestore_V1_RunQueryResponse.OneOf_ContinuationSelector?
     var _explainMetrics: Google_Firestore_V1_ExplainMetrics? = nil
 
-    static let defaultInstance = _StorageClass()
+    #if swift(>=5.10)
+      // This property is used as the initial default value for new instances of the type.
+      // The type itself is protecting the reference to its storage via CoW semantics.
+      // This will force a copy to be made of this reference when the first mutation occurs;
+      // hence, it is safe to mark this as `nonisolated(unsafe)`.
+      static nonisolated(unsafe) let defaultInstance = _StorageClass()
+    #else
+      static let defaultInstance = _StorageClass()
+    #endif
 
     private init() {}
 
