@@ -1,5 +1,6 @@
 // DO NOT EDIT.
 // swift-format-ignore-file
+// swiftlint:disable all
 //
 // Generated by the Swift generator plugin for the protocol buffer compiler.
 // Source: google/firestore/v1/query.proto
@@ -21,7 +22,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import Foundation
 import SwiftProtobuf
 
 // If the compiler emits an error on this type, it is because this file
@@ -43,7 +43,7 @@ fileprivate struct _GeneratedWithProtocGenSwiftVersion: SwiftProtobuf.ProtobufAP
 /// 4. order_by + start_at + end_at
 /// 5. offset
 /// 6. limit
-public struct Google_Firestore_V1_StructuredQuery {
+public struct Google_Firestore_V1_StructuredQuery: @unchecked Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -207,7 +207,7 @@ public struct Google_Firestore_V1_StructuredQuery {
   public var unknownFields = SwiftProtobuf.UnknownStorage()
 
   /// A sort direction.
-  public enum Direction: SwiftProtobuf.Enum {
+  public enum Direction: SwiftProtobuf.Enum, Swift.CaseIterable {
     public typealias RawValue = Int
 
     /// Unspecified.
@@ -242,10 +242,17 @@ public struct Google_Firestore_V1_StructuredQuery {
       }
     }
 
+    // The compiler won't synthesize support with the UNRECOGNIZED case.
+    public static let allCases: [Google_Firestore_V1_StructuredQuery.Direction] = [
+      .unspecified,
+      .ascending,
+      .descending,
+    ]
+
   }
 
   /// A selection of a collection, such as `messages as m1`.
-  public struct CollectionSelector {
+  public struct CollectionSelector: Sendable {
     // SwiftProtobuf.Message conformance is added in an extension below. See the
     // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
     // methods supported on all messages.
@@ -265,7 +272,7 @@ public struct Google_Firestore_V1_StructuredQuery {
   }
 
   /// A filter.
-  public struct Filter {
+  public struct Filter: Sendable {
     // SwiftProtobuf.Message conformance is added in an extension below. See the
     // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
     // methods supported on all messages.
@@ -303,7 +310,7 @@ public struct Google_Firestore_V1_StructuredQuery {
     public var unknownFields = SwiftProtobuf.UnknownStorage()
 
     /// The type of filter.
-    public enum OneOf_FilterType: Equatable {
+    public enum OneOf_FilterType: Equatable, Sendable {
       /// A composite filter.
       case compositeFilter(Google_Firestore_V1_StructuredQuery.CompositeFilter)
       /// A filter on a document field.
@@ -311,35 +318,13 @@ public struct Google_Firestore_V1_StructuredQuery {
       /// A filter that takes exactly one argument.
       case unaryFilter(Google_Firestore_V1_StructuredQuery.UnaryFilter)
 
-    #if !swift(>=4.1)
-      public static func ==(lhs: Google_Firestore_V1_StructuredQuery.Filter.OneOf_FilterType, rhs: Google_Firestore_V1_StructuredQuery.Filter.OneOf_FilterType) -> Bool {
-        // The use of inline closures is to circumvent an issue where the compiler
-        // allocates stack space for every case branch when no optimizations are
-        // enabled. https://github.com/apple/swift-protobuf/issues/1034
-        switch (lhs, rhs) {
-        case (.compositeFilter, .compositeFilter): return {
-          guard case .compositeFilter(let l) = lhs, case .compositeFilter(let r) = rhs else { preconditionFailure() }
-          return l == r
-        }()
-        case (.fieldFilter, .fieldFilter): return {
-          guard case .fieldFilter(let l) = lhs, case .fieldFilter(let r) = rhs else { preconditionFailure() }
-          return l == r
-        }()
-        case (.unaryFilter, .unaryFilter): return {
-          guard case .unaryFilter(let l) = lhs, case .unaryFilter(let r) = rhs else { preconditionFailure() }
-          return l == r
-        }()
-        default: return false
-        }
-      }
-    #endif
     }
 
     public init() {}
   }
 
   /// A filter that merges multiple other filters using the given operator.
-  public struct CompositeFilter {
+  public struct CompositeFilter: Sendable {
     // SwiftProtobuf.Message conformance is added in an extension below. See the
     // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
     // methods supported on all messages.
@@ -357,7 +342,7 @@ public struct Google_Firestore_V1_StructuredQuery {
     public var unknownFields = SwiftProtobuf.UnknownStorage()
 
     /// A composite filter operator.
-    public enum Operator: SwiftProtobuf.Enum {
+    public enum Operator: SwiftProtobuf.Enum, Swift.CaseIterable {
       public typealias RawValue = Int
 
       /// Unspecified. This value must not be used.
@@ -392,13 +377,20 @@ public struct Google_Firestore_V1_StructuredQuery {
         }
       }
 
+      // The compiler won't synthesize support with the UNRECOGNIZED case.
+      public static let allCases: [Google_Firestore_V1_StructuredQuery.CompositeFilter.Operator] = [
+        .unspecified,
+        .and,
+        .or,
+      ]
+
     }
 
     public init() {}
   }
 
   /// A filter on a specific field.
-  public struct FieldFilter {
+  public struct FieldFilter: Sendable {
     // SwiftProtobuf.Message conformance is added in an extension below. See the
     // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
     // methods supported on all messages.
@@ -429,7 +421,7 @@ public struct Google_Firestore_V1_StructuredQuery {
     public var unknownFields = SwiftProtobuf.UnknownStorage()
 
     /// A field filter operator.
-    public enum Operator: SwiftProtobuf.Enum {
+    public enum Operator: SwiftProtobuf.Enum, Swift.CaseIterable {
       public typealias RawValue = Int
 
       /// Unspecified. This value must not be used.
@@ -546,6 +538,21 @@ public struct Google_Firestore_V1_StructuredQuery {
         }
       }
 
+      // The compiler won't synthesize support with the UNRECOGNIZED case.
+      public static let allCases: [Google_Firestore_V1_StructuredQuery.FieldFilter.Operator] = [
+        .unspecified,
+        .lessThan,
+        .lessThanOrEqual,
+        .greaterThan,
+        .greaterThanOrEqual,
+        .equal,
+        .notEqual,
+        .arrayContains,
+        .in,
+        .arrayContainsAny,
+        .notIn,
+      ]
+
     }
 
     public init() {}
@@ -555,7 +562,7 @@ public struct Google_Firestore_V1_StructuredQuery {
   }
 
   /// A filter with a single operand.
-  public struct UnaryFilter {
+  public struct UnaryFilter: Sendable {
     // SwiftProtobuf.Message conformance is added in an extension below. See the
     // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
     // methods supported on all messages.
@@ -578,27 +585,14 @@ public struct Google_Firestore_V1_StructuredQuery {
     public var unknownFields = SwiftProtobuf.UnknownStorage()
 
     /// The argument to the filter.
-    public enum OneOf_OperandType: Equatable {
+    public enum OneOf_OperandType: Equatable, Sendable {
       /// The field to which to apply the operator.
       case field(Google_Firestore_V1_StructuredQuery.FieldReference)
 
-    #if !swift(>=4.1)
-      public static func ==(lhs: Google_Firestore_V1_StructuredQuery.UnaryFilter.OneOf_OperandType, rhs: Google_Firestore_V1_StructuredQuery.UnaryFilter.OneOf_OperandType) -> Bool {
-        // The use of inline closures is to circumvent an issue where the compiler
-        // allocates stack space for every case branch when no optimizations are
-        // enabled. https://github.com/apple/swift-protobuf/issues/1034
-        switch (lhs, rhs) {
-        case (.field, .field): return {
-          guard case .field(let l) = lhs, case .field(let r) = rhs else { preconditionFailure() }
-          return l == r
-        }()
-        }
-      }
-    #endif
     }
 
     /// A unary operator.
-    public enum Operator: SwiftProtobuf.Enum {
+    public enum Operator: SwiftProtobuf.Enum, Swift.CaseIterable {
       public typealias RawValue = Int
 
       /// Unspecified. This value must not be used.
@@ -653,13 +647,22 @@ public struct Google_Firestore_V1_StructuredQuery {
         }
       }
 
+      // The compiler won't synthesize support with the UNRECOGNIZED case.
+      public static let allCases: [Google_Firestore_V1_StructuredQuery.UnaryFilter.Operator] = [
+        .unspecified,
+        .isNan,
+        .isNull,
+        .isNotNan,
+        .isNotNull,
+      ]
+
     }
 
     public init() {}
   }
 
   /// An order on a field.
-  public struct Order {
+  public struct Order: Sendable {
     // SwiftProtobuf.Message conformance is added in an extension below. See the
     // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
     // methods supported on all messages.
@@ -685,7 +688,7 @@ public struct Google_Firestore_V1_StructuredQuery {
   }
 
   /// A reference to a field in a document, ex: `stats.operations`.
-  public struct FieldReference {
+  public struct FieldReference: Sendable {
     // SwiftProtobuf.Message conformance is added in an extension below. See the
     // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
     // methods supported on all messages.
@@ -705,7 +708,7 @@ public struct Google_Firestore_V1_StructuredQuery {
   }
 
   /// The projection of document's fields to return.
-  public struct Projection {
+  public struct Projection: Sendable {
     // SwiftProtobuf.Message conformance is added in an extension below. See the
     // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
     // methods supported on all messages.
@@ -725,7 +728,7 @@ public struct Google_Firestore_V1_StructuredQuery {
   /// supersedes the order_by stage. If multiple documents have the same vector
   /// distance, the returned document order is not guaranteed to be stable
   /// between queries.
-  public struct FindNearest {
+  public struct FindNearest: Sendable {
     // SwiftProtobuf.Message conformance is added in an extension below. See the
     // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
     // methods supported on all messages.
@@ -792,7 +795,7 @@ public struct Google_Firestore_V1_StructuredQuery {
     public var unknownFields = SwiftProtobuf.UnknownStorage()
 
     /// The distance measure to use when comparing vectors.
-    public enum DistanceMeasure: SwiftProtobuf.Enum {
+    public enum DistanceMeasure: SwiftProtobuf.Enum, Swift.CaseIterable {
       public typealias RawValue = Int
 
       /// Should not be set.
@@ -844,6 +847,14 @@ public struct Google_Firestore_V1_StructuredQuery {
         }
       }
 
+      // The compiler won't synthesize support with the UNRECOGNIZED case.
+      public static let allCases: [Google_Firestore_V1_StructuredQuery.FindNearest.DistanceMeasure] = [
+        .unspecified,
+        .euclidean,
+        .cosine,
+        .dotProduct,
+      ]
+
     }
 
     public init() {}
@@ -859,69 +870,9 @@ public struct Google_Firestore_V1_StructuredQuery {
   fileprivate var _storage = _StorageClass.defaultInstance
 }
 
-#if swift(>=4.2)
-
-extension Google_Firestore_V1_StructuredQuery.Direction: CaseIterable {
-  // The compiler won't synthesize support with the UNRECOGNIZED case.
-  public static let allCases: [Google_Firestore_V1_StructuredQuery.Direction] = [
-    .unspecified,
-    .ascending,
-    .descending,
-  ]
-}
-
-extension Google_Firestore_V1_StructuredQuery.CompositeFilter.Operator: CaseIterable {
-  // The compiler won't synthesize support with the UNRECOGNIZED case.
-  public static let allCases: [Google_Firestore_V1_StructuredQuery.CompositeFilter.Operator] = [
-    .unspecified,
-    .and,
-    .or,
-  ]
-}
-
-extension Google_Firestore_V1_StructuredQuery.FieldFilter.Operator: CaseIterable {
-  // The compiler won't synthesize support with the UNRECOGNIZED case.
-  public static let allCases: [Google_Firestore_V1_StructuredQuery.FieldFilter.Operator] = [
-    .unspecified,
-    .lessThan,
-    .lessThanOrEqual,
-    .greaterThan,
-    .greaterThanOrEqual,
-    .equal,
-    .notEqual,
-    .arrayContains,
-    .in,
-    .arrayContainsAny,
-    .notIn,
-  ]
-}
-
-extension Google_Firestore_V1_StructuredQuery.UnaryFilter.Operator: CaseIterable {
-  // The compiler won't synthesize support with the UNRECOGNIZED case.
-  public static let allCases: [Google_Firestore_V1_StructuredQuery.UnaryFilter.Operator] = [
-    .unspecified,
-    .isNan,
-    .isNull,
-    .isNotNan,
-    .isNotNull,
-  ]
-}
-
-extension Google_Firestore_V1_StructuredQuery.FindNearest.DistanceMeasure: CaseIterable {
-  // The compiler won't synthesize support with the UNRECOGNIZED case.
-  public static let allCases: [Google_Firestore_V1_StructuredQuery.FindNearest.DistanceMeasure] = [
-    .unspecified,
-    .euclidean,
-    .cosine,
-    .dotProduct,
-  ]
-}
-
-#endif  // swift(>=4.2)
-
 /// Firestore query for running an aggregation over a
 /// [StructuredQuery][google.firestore.v1.StructuredQuery].
-public struct Google_Firestore_V1_StructuredAggregationQuery {
+public struct Google_Firestore_V1_StructuredAggregationQuery: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -949,27 +900,14 @@ public struct Google_Firestore_V1_StructuredAggregationQuery {
   public var unknownFields = SwiftProtobuf.UnknownStorage()
 
   /// The base query to aggregate over.
-  public enum OneOf_QueryType: Equatable {
+  public enum OneOf_QueryType: Equatable, Sendable {
     /// Nested structured query.
     case structuredQuery(Google_Firestore_V1_StructuredQuery)
 
-  #if !swift(>=4.1)
-    public static func ==(lhs: Google_Firestore_V1_StructuredAggregationQuery.OneOf_QueryType, rhs: Google_Firestore_V1_StructuredAggregationQuery.OneOf_QueryType) -> Bool {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
-      switch (lhs, rhs) {
-      case (.structuredQuery, .structuredQuery): return {
-        guard case .structuredQuery(let l) = lhs, case .structuredQuery(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      }
-    }
-  #endif
   }
 
   /// Defines an aggregation that produces a single result.
-  public struct Aggregation {
+  public struct Aggregation: Sendable {
     // SwiftProtobuf.Message conformance is added in an extension below. See the
     // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
     // methods supported on all messages.
@@ -1044,7 +982,7 @@ public struct Google_Firestore_V1_StructuredAggregationQuery {
     public var unknownFields = SwiftProtobuf.UnknownStorage()
 
     /// The type of aggregation to perform, required.
-    public enum OneOf_Operator: Equatable {
+    public enum OneOf_Operator: Equatable, Sendable {
       /// Count aggregator.
       case count(Google_Firestore_V1_StructuredAggregationQuery.Aggregation.Count)
       /// Sum aggregator.
@@ -1052,35 +990,13 @@ public struct Google_Firestore_V1_StructuredAggregationQuery {
       /// Average aggregator.
       case avg(Google_Firestore_V1_StructuredAggregationQuery.Aggregation.Avg)
 
-    #if !swift(>=4.1)
-      public static func ==(lhs: Google_Firestore_V1_StructuredAggregationQuery.Aggregation.OneOf_Operator, rhs: Google_Firestore_V1_StructuredAggregationQuery.Aggregation.OneOf_Operator) -> Bool {
-        // The use of inline closures is to circumvent an issue where the compiler
-        // allocates stack space for every case branch when no optimizations are
-        // enabled. https://github.com/apple/swift-protobuf/issues/1034
-        switch (lhs, rhs) {
-        case (.count, .count): return {
-          guard case .count(let l) = lhs, case .count(let r) = rhs else { preconditionFailure() }
-          return l == r
-        }()
-        case (.sum, .sum): return {
-          guard case .sum(let l) = lhs, case .sum(let r) = rhs else { preconditionFailure() }
-          return l == r
-        }()
-        case (.avg, .avg): return {
-          guard case .avg(let l) = lhs, case .avg(let r) = rhs else { preconditionFailure() }
-          return l == r
-        }()
-        default: return false
-        }
-      }
-    #endif
     }
 
     /// Count of documents that match the query.
     ///
     /// The `COUNT(*)` aggregation function operates on the entire document
     /// so it does not require a field reference.
-    public struct Count {
+    public struct Count: Sendable {
       // SwiftProtobuf.Message conformance is added in an extension below. See the
       // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
       // methods supported on all messages.
@@ -1139,7 +1055,7 @@ public struct Google_Firestore_V1_StructuredAggregationQuery {
     /// the underlying values could produce slightly different results each
     /// time. In those cases, values should be stored as integers over
     /// floating-point numbers.
-    public struct Sum {
+    public struct Sum: Sendable {
       // SwiftProtobuf.Message conformance is added in an extension below. See the
       // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
       // methods supported on all messages.
@@ -1172,7 +1088,7 @@ public struct Google_Firestore_V1_StructuredAggregationQuery {
     /// * If the aggregated value set is empty, returns `NULL`.
     ///
     /// * Always returns the result as a double.
-    public struct Avg {
+    public struct Avg: Sendable {
       // SwiftProtobuf.Message conformance is added in an extension below. See the
       // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
       // methods supported on all messages.
@@ -1201,7 +1117,7 @@ public struct Google_Firestore_V1_StructuredAggregationQuery {
 }
 
 /// A position in a query result set.
-public struct Google_Firestore_V1_Cursor {
+public struct Google_Firestore_V1_Cursor: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -1220,34 +1136,6 @@ public struct Google_Firestore_V1_Cursor {
 
   public init() {}
 }
-
-#if swift(>=5.5) && canImport(_Concurrency)
-extension Google_Firestore_V1_StructuredQuery: @unchecked Sendable {}
-extension Google_Firestore_V1_StructuredQuery.Direction: @unchecked Sendable {}
-extension Google_Firestore_V1_StructuredQuery.CollectionSelector: @unchecked Sendable {}
-extension Google_Firestore_V1_StructuredQuery.Filter: @unchecked Sendable {}
-extension Google_Firestore_V1_StructuredQuery.Filter.OneOf_FilterType: @unchecked Sendable {}
-extension Google_Firestore_V1_StructuredQuery.CompositeFilter: @unchecked Sendable {}
-extension Google_Firestore_V1_StructuredQuery.CompositeFilter.Operator: @unchecked Sendable {}
-extension Google_Firestore_V1_StructuredQuery.FieldFilter: @unchecked Sendable {}
-extension Google_Firestore_V1_StructuredQuery.FieldFilter.Operator: @unchecked Sendable {}
-extension Google_Firestore_V1_StructuredQuery.UnaryFilter: @unchecked Sendable {}
-extension Google_Firestore_V1_StructuredQuery.UnaryFilter.OneOf_OperandType: @unchecked Sendable {}
-extension Google_Firestore_V1_StructuredQuery.UnaryFilter.Operator: @unchecked Sendable {}
-extension Google_Firestore_V1_StructuredQuery.Order: @unchecked Sendable {}
-extension Google_Firestore_V1_StructuredQuery.FieldReference: @unchecked Sendable {}
-extension Google_Firestore_V1_StructuredQuery.Projection: @unchecked Sendable {}
-extension Google_Firestore_V1_StructuredQuery.FindNearest: @unchecked Sendable {}
-extension Google_Firestore_V1_StructuredQuery.FindNearest.DistanceMeasure: @unchecked Sendable {}
-extension Google_Firestore_V1_StructuredAggregationQuery: @unchecked Sendable {}
-extension Google_Firestore_V1_StructuredAggregationQuery.OneOf_QueryType: @unchecked Sendable {}
-extension Google_Firestore_V1_StructuredAggregationQuery.Aggregation: @unchecked Sendable {}
-extension Google_Firestore_V1_StructuredAggregationQuery.Aggregation.OneOf_Operator: @unchecked Sendable {}
-extension Google_Firestore_V1_StructuredAggregationQuery.Aggregation.Count: @unchecked Sendable {}
-extension Google_Firestore_V1_StructuredAggregationQuery.Aggregation.Sum: @unchecked Sendable {}
-extension Google_Firestore_V1_StructuredAggregationQuery.Aggregation.Avg: @unchecked Sendable {}
-extension Google_Firestore_V1_Cursor: @unchecked Sendable {}
-#endif  // swift(>=5.5) && canImport(_Concurrency)
 
 // MARK: - Code below here is support for the SwiftProtobuf runtime.
 
@@ -1278,7 +1166,15 @@ extension Google_Firestore_V1_StructuredQuery: SwiftProtobuf.Message, SwiftProto
     var _limit: SwiftProtobuf.Google_Protobuf_Int32Value? = nil
     var _findNearest: Google_Firestore_V1_StructuredQuery.FindNearest? = nil
 
-    static let defaultInstance = _StorageClass()
+    #if swift(>=5.10)
+      // This property is used as the initial default value for new instances of the type.
+      // The type itself is protecting the reference to its storage via CoW semantics.
+      // This will force a copy to be made of this reference when the first mutation occurs;
+      // hence, it is safe to mark this as `nonisolated(unsafe)`.
+      static nonisolated(unsafe) let defaultInstance = _StorageClass()
+    #else
+      static let defaultInstance = _StorageClass()
+    #endif
 
     private init() {}
 
