@@ -1,5 +1,6 @@
 // DO NOT EDIT.
 // swift-format-ignore-file
+// swiftlint:disable all
 //
 // Generated by the Swift generator plugin for the protocol buffer compiler.
 // Source: google/firestore/v1/query_profile.proto
@@ -21,7 +22,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import Foundation
 import SwiftProtobuf
 
 // If the compiler emits an error on this type, it is because this file
@@ -35,7 +35,7 @@ fileprivate struct _GeneratedWithProtocGenSwiftVersion: SwiftProtobuf.ProtobufAP
 }
 
 /// Explain options for the query.
-public struct Google_Firestore_V1_ExplainOptions {
+public struct Google_Firestore_V1_ExplainOptions: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -55,7 +55,7 @@ public struct Google_Firestore_V1_ExplainOptions {
 }
 
 /// Explain metrics for the query.
-public struct Google_Firestore_V1_ExplainMetrics {
+public struct Google_Firestore_V1_ExplainMetrics: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -91,7 +91,7 @@ public struct Google_Firestore_V1_ExplainMetrics {
 }
 
 /// Planning phase information for the query.
-public struct Google_Firestore_V1_PlanSummary {
+public struct Google_Firestore_V1_PlanSummary: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -109,7 +109,7 @@ public struct Google_Firestore_V1_PlanSummary {
 }
 
 /// Execution statistics for the query.
-public struct Google_Firestore_V1_ExecutionStats {
+public struct Google_Firestore_V1_ExecutionStats: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -159,13 +159,6 @@ public struct Google_Firestore_V1_ExecutionStats {
   fileprivate var _executionDuration: SwiftProtobuf.Google_Protobuf_Duration? = nil
   fileprivate var _debugStats: SwiftProtobuf.Google_Protobuf_Struct? = nil
 }
-
-#if swift(>=5.5) && canImport(_Concurrency)
-extension Google_Firestore_V1_ExplainOptions: @unchecked Sendable {}
-extension Google_Firestore_V1_ExplainMetrics: @unchecked Sendable {}
-extension Google_Firestore_V1_PlanSummary: @unchecked Sendable {}
-extension Google_Firestore_V1_ExecutionStats: @unchecked Sendable {}
-#endif  // swift(>=5.5) && canImport(_Concurrency)
 
 // MARK: - Code below here is support for the SwiftProtobuf runtime.
 
