@@ -1,5 +1,6 @@
 // DO NOT EDIT.
 // swift-format-ignore-file
+// swiftlint:disable all
 //
 // Generated by the Swift generator plugin for the protocol buffer compiler.
 // Source: google/firestore/v1/write.proto
@@ -21,7 +22,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import Foundation
 import SwiftProtobuf
 
 // If the compiler emits an error on this type, it is because this file
@@ -35,7 +35,7 @@ fileprivate struct _GeneratedWithProtocGenSwiftVersion: SwiftProtobuf.ProtobufAP
 }
 
 /// A write on a document.
-public struct Google_Firestore_V1_Write {
+public struct Google_Firestore_V1_Write: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -112,7 +112,7 @@ public struct Google_Firestore_V1_Write {
   public var unknownFields = SwiftProtobuf.UnknownStorage()
 
   /// The operation to execute.
-  public enum OneOf_Operation: Equatable {
+  public enum OneOf_Operation: Equatable, Sendable {
     /// A document to write.
     case update(Google_Firestore_V1_Document)
     /// A document name to delete. In the format:
@@ -121,28 +121,6 @@ public struct Google_Firestore_V1_Write {
     /// Applies a transformation to a document.
     case transform(Google_Firestore_V1_DocumentTransform)
 
-  #if !swift(>=4.1)
-    public static func ==(lhs: Google_Firestore_V1_Write.OneOf_Operation, rhs: Google_Firestore_V1_Write.OneOf_Operation) -> Bool {
-      // The use of inline closures is to circumvent an issue where the compiler
-      // allocates stack space for every case branch when no optimizations are
-      // enabled. https://github.com/apple/swift-protobuf/issues/1034
-      switch (lhs, rhs) {
-      case (.update, .update): return {
-        guard case .update(let l) = lhs, case .update(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      case (.delete, .delete): return {
-        guard case .delete(let l) = lhs, case .delete(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      case (.transform, .transform): return {
-        guard case .transform(let l) = lhs, case .transform(let r) = rhs else { preconditionFailure() }
-        return l == r
-      }()
-      default: return false
-      }
-    }
-  #endif
   }
 
   public init() {}
@@ -152,7 +130,7 @@ public struct Google_Firestore_V1_Write {
 }
 
 /// A transformation of a document.
-public struct Google_Firestore_V1_DocumentTransform {
+public struct Google_Firestore_V1_DocumentTransform: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -168,7 +146,7 @@ public struct Google_Firestore_V1_DocumentTransform {
   public var unknownFields = SwiftProtobuf.UnknownStorage()
 
   /// A transformation of a field of the document.
-  public struct FieldTransform {
+  public struct FieldTransform: Sendable {
     // SwiftProtobuf.Message conformance is added in an extension below. See the
     // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
     // methods supported on all messages.
@@ -289,7 +267,7 @@ public struct Google_Firestore_V1_DocumentTransform {
     public var unknownFields = SwiftProtobuf.UnknownStorage()
 
     /// The transformation to apply on the field.
-    public enum OneOf_TransformType: Equatable {
+    public enum OneOf_TransformType: Equatable, Sendable {
       /// Sets the field to the given server value.
       case setToServerValue(Google_Firestore_V1_DocumentTransform.FieldTransform.ServerValue)
       /// Adds the given value to the field's current value.
@@ -354,44 +332,10 @@ public struct Google_Firestore_V1_DocumentTransform {
       /// The corresponding transform_result will be the null value.
       case removeAllFromArray(Google_Firestore_V1_ArrayValue)
 
-    #if !swift(>=4.1)
-      public static func ==(lhs: Google_Firestore_V1_DocumentTransform.FieldTransform.OneOf_TransformType, rhs: Google_Firestore_V1_DocumentTransform.FieldTransform.OneOf_TransformType) -> Bool {
-        // The use of inline closures is to circumvent an issue where the compiler
-        // allocates stack space for every case branch when no optimizations are
-        // enabled. https://github.com/apple/swift-protobuf/issues/1034
-        switch (lhs, rhs) {
-        case (.setToServerValue, .setToServerValue): return {
-          guard case .setToServerValue(let l) = lhs, case .setToServerValue(let r) = rhs else { preconditionFailure() }
-          return l == r
-        }()
-        case (.increment, .increment): return {
-          guard case .increment(let l) = lhs, case .increment(let r) = rhs else { preconditionFailure() }
-          return l == r
-        }()
-        case (.maximum, .maximum): return {
-          guard case .maximum(let l) = lhs, case .maximum(let r) = rhs else { preconditionFailure() }
-          return l == r
-        }()
-        case (.minimum, .minimum): return {
-          guard case .minimum(let l) = lhs, case .minimum(let r) = rhs else { preconditionFailure() }
-          return l == r
-        }()
-        case (.appendMissingElements, .appendMissingElements): return {
-          guard case .appendMissingElements(let l) = lhs, case .appendMissingElements(let r) = rhs else { preconditionFailure() }
-          return l == r
-        }()
-        case (.removeAllFromArray, .removeAllFromArray): return {
-          guard case .removeAllFromArray(let l) = lhs, case .removeAllFromArray(let r) = rhs else { preconditionFailure() }
-          return l == r
-        }()
-        default: return false
-        }
-      }
-    #endif
     }
 
     /// A value that is calculated by the server.
-    public enum ServerValue: SwiftProtobuf.Enum {
+    public enum ServerValue: SwiftProtobuf.Enum, Swift.CaseIterable {
       public typealias RawValue = Int
 
       /// Unspecified. This value must not be used.
@@ -423,6 +367,12 @@ public struct Google_Firestore_V1_DocumentTransform {
         }
       }
 
+      // The compiler won't synthesize support with the UNRECOGNIZED case.
+      public static let allCases: [Google_Firestore_V1_DocumentTransform.FieldTransform.ServerValue] = [
+        .unspecified,
+        .requestTime,
+      ]
+
     }
 
     public init() {}
@@ -431,20 +381,8 @@ public struct Google_Firestore_V1_DocumentTransform {
   public init() {}
 }
 
-#if swift(>=4.2)
-
-extension Google_Firestore_V1_DocumentTransform.FieldTransform.ServerValue: CaseIterable {
-  // The compiler won't synthesize support with the UNRECOGNIZED case.
-  public static let allCases: [Google_Firestore_V1_DocumentTransform.FieldTransform.ServerValue] = [
-    .unspecified,
-    .requestTime,
-  ]
-}
-
-#endif  // swift(>=4.2)
-
 /// The result of applying a write.
-public struct Google_Firestore_V1_WriteResult {
+public struct Google_Firestore_V1_WriteResult: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -483,7 +421,7 @@ public struct Google_Firestore_V1_WriteResult {
 ///
 /// Multiple [DocumentChange][google.firestore.v1.DocumentChange] messages may be
 /// returned for the same logical change, if multiple targets are affected.
-public struct Google_Firestore_V1_DocumentChange {
+public struct Google_Firestore_V1_DocumentChange: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -521,7 +459,7 @@ public struct Google_Firestore_V1_DocumentChange {
 ///
 /// Multiple [DocumentDelete][google.firestore.v1.DocumentDelete] messages may be
 /// returned for the same logical delete, if multiple targets are affected.
-public struct Google_Firestore_V1_DocumentDelete {
+public struct Google_Firestore_V1_DocumentDelete: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -562,7 +500,7 @@ public struct Google_Firestore_V1_DocumentDelete {
 /// Multiple [DocumentRemove][google.firestore.v1.DocumentRemove] messages may be
 /// returned for the same logical write or delete, if multiple targets are
 /// affected.
-public struct Google_Firestore_V1_DocumentRemove {
+public struct Google_Firestore_V1_DocumentRemove: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -594,7 +532,7 @@ public struct Google_Firestore_V1_DocumentRemove {
 }
 
 /// A digest of all the documents that match a given target.
-public struct Google_Firestore_V1_ExistenceFilter {
+public struct Google_Firestore_V1_ExistenceFilter: Sendable {
   // SwiftProtobuf.Message conformance is added in an extension below. See the
   // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
   // methods supported on all messages.
@@ -640,20 +578,6 @@ public struct Google_Firestore_V1_ExistenceFilter {
 
   fileprivate var _unchangedNames: Google_Firestore_V1_BloomFilter? = nil
 }
-
-#if swift(>=5.5) && canImport(_Concurrency)
-extension Google_Firestore_V1_Write: @unchecked Sendable {}
-extension Google_Firestore_V1_Write.OneOf_Operation: @unchecked Sendable {}
-extension Google_Firestore_V1_DocumentTransform: @unchecked Sendable {}
-extension Google_Firestore_V1_DocumentTransform.FieldTransform: @unchecked Sendable {}
-extension Google_Firestore_V1_DocumentTransform.FieldTransform.OneOf_TransformType: @unchecked Sendable {}
-extension Google_Firestore_V1_DocumentTransform.FieldTransform.ServerValue: @unchecked Sendable {}
-extension Google_Firestore_V1_WriteResult: @unchecked Sendable {}
-extension Google_Firestore_V1_DocumentChange: @unchecked Sendable {}
-extension Google_Firestore_V1_DocumentDelete: @unchecked Sendable {}
-extension Google_Firestore_V1_DocumentRemove: @unchecked Sendable {}
-extension Google_Firestore_V1_ExistenceFilter: @unchecked Sendable {}
-#endif  // swift(>=5.5) && canImport(_Concurrency)
 
 // MARK: - Code below here is support for the SwiftProtobuf runtime.
 
