@@ -1,5 +1,6 @@
 // DO NOT EDIT.
 // swift-format-ignore-file
+// swiftlint:disable all
 //
 // Generated by the Swift generator plugin for the protocol buffer compiler.
 // Source: google/api/field_behavior.proto
@@ -21,7 +22,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import Foundation
 import SwiftProtobuf
 
 // If the compiler emits an error on this type, it is because this file
@@ -40,7 +40,7 @@ fileprivate struct _GeneratedWithProtocGenSwiftVersion: SwiftProtobuf.ProtobufAP
 /// denotes the behavior and may affect how API tooling handles the field.
 ///
 /// Note: This enum **may** receive new values in the future.
-public enum Google_Api_FieldBehavior: SwiftProtobuf.Enum {
+public enum Google_Api_FieldBehavior: SwiftProtobuf.Enum, Swift.CaseIterable {
   public typealias RawValue = Int
 
   /// Conventional default for enums. Do not use this.
@@ -133,11 +133,6 @@ public enum Google_Api_FieldBehavior: SwiftProtobuf.Enum {
     }
   }
 
-}
-
-#if swift(>=4.2)
-
-extension Google_Api_FieldBehavior: CaseIterable {
   // The compiler won't synthesize support with the UNRECOGNIZED case.
   public static let allCases: [Google_Api_FieldBehavior] = [
     .unspecified,
@@ -150,19 +145,14 @@ extension Google_Api_FieldBehavior: CaseIterable {
     .nonEmptyDefault,
     .identifier,
   ]
+
 }
-
-#endif  // swift(>=4.2)
-
-#if swift(>=5.5) && canImport(_Concurrency)
-extension Google_Api_FieldBehavior: @unchecked Sendable {}
-#endif  // swift(>=5.5) && canImport(_Concurrency)
 
 // MARK: - Extension support defined in field_behavior.proto.
 
 // MARK: - Extension Properties
 
-// Swift Extensions on the exteneded Messages to add easy access to the declared
+// Swift Extensions on the extended Messages to add easy access to the declared
 // extension fields. The names are based on the extension field name from the proto
 // declaration. To avoid naming collisions, the names are prefixed with the name of
 // the scope where the extend directive occurs.
@@ -184,16 +174,6 @@ extension SwiftProtobuf.Google_Protobuf_FieldOptions {
   public var Google_Api_fieldBehavior: [Google_Api_FieldBehavior] {
     get {return getExtensionValue(ext: Google_Api_Extensions_field_behavior) ?? []}
     set {setExtensionValue(ext: Google_Api_Extensions_field_behavior, value: newValue)}
-  }
-  /// Returns true if extension `Google_Api_Extensions_field_behavior`
-  /// has been explicitly set.
-  public var hasGoogle_Api_fieldBehavior: Bool {
-    return hasExtensionValue(ext: Google_Api_Extensions_field_behavior)
-  }
-  /// Clears the value of extension `Google_Api_Extensions_field_behavior`.
-  /// Subsequent reads from it will return its default value.
-  public mutating func clearGoogle_Api_fieldBehavior() {
-    clearExtensionValue(ext: Google_Api_Extensions_field_behavior)
   }
 
 }
